Set options to array

+++ src/main/java/com/stgsporting/piehmecup/dtos/quizzes/QuizDTO.java
@@ -14,7 +14,7 @@ public class QuizDTO {
     private Long id;
     private String name;
     private String slug;
-    private Integer coins;
+    private Long coins;
     private Boolean isSolved;
     private List<Question> questions;
