+++ src/main/java/com/stgsporting/piehmecup/entities/Quiz.java
@@ -15,7 +15,7 @@ public class Quiz {
     private Long id;
     private String name;
     private String slug;
-    private Integer coins;
+    private Long coins;
     private Boolean isSolved;
     private SchoolYear schoolYear;
     private List<Question> questions;
@@ -25,7 +25,7 @@ public class Quiz {
         quiz.setId((Long) quizJson.get("id"));
         quiz.setName((String) quizJson.get("name"));
         quiz.setSlug((String) quizJson.get("slug"));
-        quiz.setCoins((Integer) quizJson.get("points"));
+        quiz.setCoins((Long) quizJson.get("points"));
         quiz.setIsSolved((Boolean) quizJson.getOrDefault("is_solved", false));
         quiz.setSchoolYear(schoolYear);
